project/requirements.txt
Package             Version
------------------- -------
Django              2.2.6  
djangorestframework 3.10.3 
pip                 19.0.3 
pytz                2019.3 
setuptools          40.8.0 
sqlparse            0.3.0  
